+++ StreamChatUITestsAppUITests/Tests/Message Delivery Status/MessageDeliveryStatus_Tests.swift
@@ -361,8 +361,6 @@ extension MessageDeliveryStatus_Tests {
     func test_noDoubleCheckmarkShownInThreadReply_whenNewParticipantAdded() throws {
         linkToScenario(withId: 154)
         
-        try XCTSkipIf(ProcessInfo().operatingSystemVersion.majorVersion == 12, "Flaky on iOS 12")
-
         GIVEN("user opens the channel") {
             userRobot
                 .login()