+++ CHANGELOG.md
@@ -3,6 +3,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 # Upcoming
 
+### 🔄 Changed
+
+# [4.48.1](https://github.com/GetStream/stream-chat-swift/releases/tag/4.48.1)
+_February 09, 2024_
+
 ## StreamChat
 ### 🐞 Fixed
 - Fix `TextLink.url` in `TextLinkDetector` not having the HTTP URL [#3023](https://github.com/GetStream/stream-chat-swift/pull/3023)